techmate_dash/google_sheets_api_sync/requirements.txt
google_api_python_client==2.105.0
google_auth_oauthlib==1.1.0
protobuf==4.24.4
